# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Herpa Agentic GraphRAG Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/herbal-recommendations/analyze": {
      "post": {
        "tags": [
          "Herbal Recommendations"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_herbal_recommendations_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HerbalRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HerbalRecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "Herbal Recommendations"
        ],
        "summary": "History",
        "operationId": "history_api_v1_recommendations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/recommendations/{session_id}": {
      "get": {
        "tags": [
          "Herbal Recommendations"
        ],
        "summary": "Get Recommendation",
        "operationId": "get_recommendation_api_v1_recommendations__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Herbal Recommendations"
        ],
        "summary": "Delete Recommendation",
        "operationId": "delete_recommendation_api_v1_recommendations__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/storage/summary": {
      "get": {
        "tags": [
          "Admin Storage"
        ],
        "summary": "Summary",
        "operationId": "summary_api_v1_admin_storage_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Summary Api V1 Admin Storage Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/storage/objects": {
      "get": {
        "tags": [
          "Admin Storage"
        ],
        "summary": "Objects",
        "operationId": "objects_api_v1_admin_storage_objects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Objects Api V1 Admin Storage Objects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/storage/objects/{object_id}": {
      "delete": {
        "tags": [
          "Admin Storage"
        ],
        "summary": "Delete Object",
        "operationId": "delete_object_api_v1_admin_storage_objects__object_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/storage/cleanup-orphans": {
      "post": {
        "tags": [
          "Admin Storage"
        ],
        "summary": "Cleanup Orphans",
        "operationId": "cleanup_orphans_api_v1_admin_storage_cleanup_orphans_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Cleanup Orphans Api V1 Admin Storage Cleanup Orphans Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HerbalCandidate": {
        "properties": {
          "plant_id": {
            "type": "string",
            "title": "Plant Id"
          },
          "local_name": {
            "type": "string",
            "title": "Local Name"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "plant_part": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Part"
          },
          "evidence_level": {
            "type": "string",
            "title": "Evidence Level",
            "default": "unknown"
          },
          "traditional_use": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traditional Use"
          },
          "preparation_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation Note"
          },
          "contraindications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contraindications"
          },
          "drug_interactions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Drug Interactions"
          },
          "side_effects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Side Effects"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "Sources"
          },
          "field_evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Field Evidence"
          }
        },
        "type": "object",
        "required": [
          "plant_id",
          "local_name",
          "scientific_name",
          "relevance_score",
          "reason"
        ],
        "title": "HerbalCandidate"
      },
      "HerbalRecommendationRequest": {
        "properties": {
          "symptoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Symptoms"
          },
          "free_text": {
            "type": "string",
            "maxLength": 5000,
            "title": "Free Text",
            "default": ""
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ringan",
                  "sedang",
                  "berat"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "age_group": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "child",
                  "adolescent",
                  "adult",
                  "elderly"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Group"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "female",
                  "male",
                  "other",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "pregnant": {
            "type": "boolean",
            "title": "Pregnant",
            "default": false
          },
          "breastfeeding": {
            "type": "boolean",
            "title": "Breastfeeding",
            "default": false
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies"
          },
          "current_medications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Current Medications"
          },
          "medical_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Medical Conditions"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          }
        },
        "type": "object",
        "title": "HerbalRecommendationRequest"
      },
      "HerbalRecommendationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/HerbalCandidate"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/HerbalCandidate"
            },
            "type": "array",
            "title": "Options"
          },
          "excluded_candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Excluded Candidates"
          },
          "general_disclaimer": {
            "type": "string",
            "title": "General Disclaimer",
            "default": "Informasi ini bersifat edukatif dan bukan diagnosis atau pengganti tenaga kesehatan."
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Informasi ini bersifat edukatif dan bukan diagnosis atau pengganti tenaga kesehatan."
          },
          "medical_attention_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Attention Message"
          },
          "red_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Red Flags"
          },
          "when_to_seek_medical_help": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "When To Seek Medical Help"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HerbalRecommendationResponse"
      },
...
      "SourceReference": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "evidence_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Level"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "type",
          "source_id",
          "title"
        ],
        "title": "SourceReference"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}